test: caution on our version of 'yes'
During a review of a patch, we noticed that we use our own imitation
of 'yes' with the limit of 99 lines.  It is very tempting to lift this
arbitrary limit, but the limit is there for a reason.

Add an in-code comment to prevent future developers from wasting
their time.

Signed-off-by: Junio C Hamano <gitster@pobox.com>